Fixed no-invite-spam function

+++ monitors/monitor.js
@@ -3,14 +3,15 @@
 
 module.exports = class {
 
-  static run(client, message) {
-    this.givePoints(client, message),
-      this.checkAFK(client, message),
-      // this.mentions(client, message),
-      this.antiInvite(client, message);
+  static run(client, message, level) {
+    this.givePoints(client, message, level),
+    this.checkAFK(client, message, level),
+    // this.mentions(client, message),
+    this.antiInvite(client, message, level);
   }
 
-  static givePoints(client, message) {
+  static givePoints(client, message, level) { // eslint-disable-line no-unused-vars
+    if (message.channel.type !== 'text') return;
     const settings = client.settings.get(message.guild.id);
     if (message.content.startsWith(settings.prefix) || message.channel.type !== 'text') return;
     const score = client.points.get(`${message.guild.id}-${message.author.id}`) || { points: 0, level: 0, user: message.author.id, guild: message.guild.id, daily: 1504120109 };
@@ -24,7 +25,7 @@ module.exports = class {
     client.points.set(`${message.guild.id}-${message.author.id}`, score);
   }
 
-  static checkAFK(client, message) {
+  static checkAFK(client, message, level) { // eslint-disable-line no-unused-vars
     if (!message.guild) return;
     const settings = client.settings.get(message.guild.id);
     const person = message.mentions.members.first();
@@ -38,7 +39,7 @@ module.exports = class {
 
 
   // // Not ready yet.
-  // static mentions(client, message) {
+  // static mentions(client, message, level) { // eslint-disable-line no-unused-vars
   //   if (message.mentions.users.size === 1 && message.mentions.users.first().bot || !message.guild.member(client.user).hasPermission('BAN_MEMBERS')) return;
   //   try {
   //     let entry = slowmode.get(message.author.id);
@@ -69,29 +70,23 @@ module.exports = class {
   //   }
   // }
 
-  static antiInvite(client, message) {
+  static antiInvite(client, message, level) {
+    if (level > 0) return;
     if (/(discord\.(gg|io|me|li)\/.+|discordapp\.com\/invite\/.+)/i.test(message.content)) {
       message.delete().then(() => {
         let count = 1;
         const spammer = `${message.guild.id}-${message.author.id}`;
-        const list = client.invspam.get(spammer) || client.invspam.set(spammer, {
-          count: 0
-        }).get(spammer);
+        const list = client.invspam.get(spammer) || client.invspam.set(spammer, { count: 0 }).get(spammer);
         if (list) count = list.count + 1;
         if (count >= parseInt(client.settings.get(message.guild.id).inviteLimit)) {
-          message.member.ban({
-            days: 2,
-            reason: 'Automatic ban, invite spam threshold exceeded.'
-          }).then((g) => {
+          message.member.ban({ days: 2, reason: 'Automatic ban, invite spam threshold exceeded.' }).then((g) => {
             message.channel.send(`${g.user.username} was successfully banned for invite spam`);
             client.invspam.delete(spammer);
           });
         }
-        client.invspam.set(spammer, {
-          count
-        });
+        client.invspam.set(spammer, { count });
       });
-      message.reply('your message contained a server invite link, which this server prohibits').then((m) => m.delete(5000));
+      message.channel.send(`${message.author} |\`🛑\`| Your message contained a server invite link, which this server prohibits.`);
     }
   }
 };
